Move seriveflags and make casts clearer.
