+++ src/ipc/operation/mod.rs
@@ -3,6 +3,19 @@ pub mod resolve;
 pub mod addrinfo;
 pub mod register;
 
+#[repr(u32)]
+#[derive(Copy, Clone, Debug)]
+pub enum ServiceFlags
+{
+    None = 0x0,
+    AutoTrigger = 0x1,
+    Add = 0x2,
+    Default = 0x3,
+    ForceMulticast = 0x400,
+    IncludeP2p = 0x20000,
+    IncludeAwdl = 0x100000,
+}
+
 #[derive(Debug, PartialEq, Eq)]
 pub enum ReplyFlags
 {
