+++ src/ipc/operation/resolve.rs
@@ -18,18 +18,7 @@ const ESCAPED_BYTE_LARGE: &str = "\\127\\128\\129\
      \\240\\241\\242\\243\\244\\245\\246\\247\\248\\249\
      \\250\\251\\252\\253\\254\\255";
 
-#[repr(u32)]
-#[derive(Copy, Clone, Debug)]
-pub enum ServiceFlags
-{
-    None = 0x0,
-    AutoTrigger = 0x1,
-    Add = 0x2,
-    Default = 0x3,
-    ForceMulticast = 0x400,
-    IncludeP2p = 0x20000,
-    IncludeAwdl = 0x100000,
-}
+use super::ServiceFlags;
 
 pub struct Request
 {
